rename braintree paypal button file
