+++ packages/react-paypal-js/src/v6/components/Braintree/BraintreePayPalOneTimePaymentButton.tsx
