+++ packages/react-paypal-js/src/v6/index.ts
@@ -43,7 +43,7 @@ export {
 } from "./components/PayPalSubscriptionButton";
 export { PayPalProvider } from "./components/PayPalProvider";
 export { BraintreePayPalProvider } from "./components/Braintree/BraintreePayPalProvider";
-export { BraintreePayPalOneTimePaymentButton } from "./components/Braintree/BraintreePayPalButton";
+export { BraintreePayPalOneTimePaymentButton } from "./components/Braintree/BraintreePayPalOneTimePaymentButton";
 export { PayPalSavePaymentButton } from "./components/PayPalSavePaymentButton";
 export { VenmoOneTimePaymentButton } from "./components/VenmoOneTimePaymentButton";
 export {
